Restructure the package
